Fetches all the news according to the category given in the navbar.
This version is able to fetch news according to category whose links are given in the navbar section. It makes use of react router so the transition between categories is smooth. Need to add features like fetch by country name and etc. Can definitely improve the colour scheme of the whole website.