All the CRUD and Authentication works flawlessly
All the CRUD operations gets executed perfectly and the Authentication of users also gets done correctly. Notes are shown of those person who have logged in and they can view only their notes. Just need to get the UI some upgrades and then the website will be ready to be deployed to the server.